# docs/openapi.json
{
  "components": {
    "schemas": {
      "AcademicStatus": {
        "enum": [
          "active",
          "on_leave",
          "graduated",
          "suspended"
        ],
        "title": "AcademicStatus",
        "type": "string"
      },
...
      "Body_import_students_students_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_students_students_import_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RosterImportCredential": {
        "properties": {
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "registration_number",
          "temporary_password"
        ],
        "title": "RosterImportCredential",
        "type": "object"
      },
...
      "RosterImportResponse": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "can_apply": {
            "title": "Can Apply",
            "type": "boolean"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/RosterImportCredential"
            },
            "title": "Credentials",
            "type": "array"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "duplicates": {
            "title": "Duplicates",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RosterImportError"
            },
            "title": "Errors",
            "type": "array"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "invalid": {
            "title": "Invalid",
            "type": "integer"
          },
          "rows_read": {
            "title": "Rows Read",
            "type": "integer"
          },
          "would_create": {
            "title": "Would Create",
            "type": "integer"
          },
          "would_update": {
            "title": "Would Update",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "rows_read",
          "would_create",
          "would_update",
          "duplicates",
          "invalid",
          "can_apply",
          "applied",
          "dry_run",
          "errors"
        ],
        "title": "RosterImportResponse",
        "type": "object"
      },
...
      "StudentIdentityListResponse": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "students": {
            "items": {
              "$ref": "#/components/schemas/StudentIdentityResponse"
            },
            "title": "Students",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "students",
          "total",
          "offset",
          "limit"
        ],
        "title": "StudentIdentityListResponse",
        "type": "object"
      },
      "StudentIdentityResponse": {
        "properties": {
          "academic_status": {
            "$ref": "#/components/schemas/AcademicStatus"
          },
          "batch": {
            "title": "Batch",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_semester": {
            "title": "Current Semester",
            "type": "integer"
          },
          "department": {
            "title": "Department",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "institutional_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutional Email"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "program": {
            "title": "Program",
            "type": "string"
          },
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "registration_number",
          "full_name",
          "institutional_email",
          "department",
          "program",
          "batch",
          "current_semester",
          "section",
          "academic_status",
          "is_verified",
          "is_active",
          "must_change_password",
          "preferred_name",
          "onboarding_completed",
          "created_at",
          "updated_at"
        ],
        "title": "StudentIdentityResponse",
        "type": "object"
      },
...
      "StudentOnboardingUpdate": {
        "additionalProperties": false,
        "properties": {
          "complete_onboarding": {
            "const": true,
            "title": "Complete Onboarding",
            "type": "boolean"
          },
          "preferred_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          }
        },
        "required": [
          "complete_onboarding"
        ],
        "title": "StudentOnboardingUpdate",
        "type": "object"
      },
      "StudentProvisionCreate": {
        "additionalProperties": false,
        "properties": {
          "academic_status": {
            "$ref": "#/components/schemas/AcademicStatus",
            "default": "active"
          },
          "batch": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Batch",
            "type": "string"
          },
          "current_semester": {
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Current Semester",
            "type": "integer"
          },
          "department": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Department",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Full Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "is_verified": {
            "default": true,
            "title": "Is Verified",
            "type": "boolean"
          },
          "program": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Program",
            "type": "string"
          },
          "registration_number": {
            "maxLength": 50,
            "minLength": 3,
            "title": "Registration Number",
            "type": "string"
          },
          "section": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Section",
            "type": "string"
          },
          "temporary_password": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporary Password"
          }
        },
        "required": [
          "registration_number",
          "full_name",
          "department",
          "program",
          "batch",
          "current_semester",
          "section"
        ],
        "title": "StudentProvisionCreate",
        "type": "object"
      },
      "StudentProvisionResponse": {
        "properties": {
          "student": {
            "$ref": "#/components/schemas/StudentIdentityResponse"
          },
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "student",
          "temporary_password"
        ],
        "title": "StudentProvisionResponse",
        "type": "object"
      },
...
      "TemporaryCredentialResponse": {
        "properties": {
          "must_change_password": {
            "const": true,
            "default": true,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "registration_number": {
            "title": "Registration Number",
            "type": "string"
          },
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          }
        },
        "required": [
          "registration_number",
          "temporary_password"
        ],
        "title": "TemporaryCredentialResponse",
        "type": "object"
      },
...
      "TimetableEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "class_type": {
            "default": "lecture",
            "enum": [
              "lecture",
              "lab",
              "tutorial",
              "online",
              "hybrid",
              "other"
            ],
            "title": "Class Type",
            "type": "string"
          },
          "course_code": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "day": {
            "enum": [
              "Monday",
              "Tuesday",
              "Wednesday",
              "Thursday",
              "Friday",
              "Saturday",
              "Sunday"
            ],
            "title": "Day",
            "type": "string"
          },
          "end_time": {
            "title": "End Time",
            "type": "string"
          },
          "entry_kind": {
            "default": "course",
            "enum": [
              "course",
              "special_event"
            ],
            "title": "Entry Kind",
            "type": "string"
          },
          "faculty": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "raw_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "room": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "section": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "semester": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semester"
          },
          "source": {
            "default": "manual",
            "enum": [
              "manual",
              "csv",
              "xlsx",
              "docx",
              "pdf",
              "image"
            ],
            "title": "Source",
            "type": "string"
          },
          "start_time": {
            "title": "Start Time",
            "type": "string"
          },
          "term_id": {
            "title": "Term Id",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "start_time",
          "end_time",
          "id",
          "term_id"
        ],
        "title": "TimetableEntryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "UniTime AI API",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/account/student-profile": {
      "get": {
        "operationId": "get_my_student_profile_account_student_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Student Profile",
        "tags": [
          "Account"
        ]
      },
      "patch": {
        "operationId": "complete_my_student_profile_account_student_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentOnboardingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete My Student Profile",
        "tags": [
          "Account"
        ]
      }
    },
...
    "/student/timetable": {
      "get": {
        "operationId": "get_my_timetable_student_timetable_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimetableEntryResponse"
                  },
                  "title": "Response Get My Timetable Student Timetable Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Timetable",
        "tags": [
          "Student"
        ]
      }
    },
    "/students": {
      "get": {
        "operationId": "get_students_students_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "is_verified",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Verified"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentIdentityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Students",
        "tags": [
          "Student Provisioning"
        ]
      },
      "post": {
        "operationId": "create_student_students_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentProvisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProvisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Student",
        "tags": [
          "Student Provisioning"
        ]
      }
    },
    "/students/import": {
      "post": {
        "operationId": "import_students_students_import_post",
        "parameters": [
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "update_existing",
            "required": false,
            "schema": {
              "default": false,
              "title": "Update Existing",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_students_students_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Students",
        "tags": [
          "Student Provisioning"
        ]
      }
    },
    "/students/{user_id}": {
      "get": {
        "operationId": "get_student_students__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Student",
        "tags": [
          "Student Provisioning"
        ]
      },
      "patch": {
        "operationId": "update_student_students__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentIdentityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentIdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Student",
        "tags": [
          "Student Provisioning"
        ]
      }
    },
    "/students/{user_id}/temporary-password": {
      "post": {
        "operationId": "reset_temporary_password_students__user_id__temporary_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemporaryCredentialResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Temporary Password",
        "tags": [
          "Student Provisioning"
        ]
      }
    },
    "/timetable": {
      "get": {
        "operationId": "get_timetable_timetable_get",
        "parameters": [
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimetableEntryResponse"
                  },
                  "title": "Response Get Timetable Timetable Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Timetable"
      },
      "post": {
        "operationId": "create_timetable_entry_timetable_post",
        "parameters": [
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableEntryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Timetable Entry"
      }
    },
...
  }
}